Move common logic to base classes
